feat: 방향키 설정 및 Screen.tsx 분리
- redUser/blueUser 확인 후 camp 설정
- 게임화면 Screen.tsx 로 분리

#104
